refactor organization
