+++ mol_gnn/featurizers/graph/__init__.py
@@ -1,4 +1,4 @@
 from .base import GraphFeaturizer
-from .graph import Graph
+from ...data.models.graph import Graph
 from .molecule import MolGraphFeaturizer
 from .reaction import CondensedReactionGraphFeaturizer, CGRFeaturizer, RxnMode
